requirements.txt
# ── Backend ───────────────────────────────────────────────
fastapi
uvicorn[standard]
python-dotenv
httpx

# ── Gemini ────────────────────────────────────────────────
google-genai

# ── Audio ─────────────────────────────────────────────────
pyaudio
numpy

# ── Document parsing ──────────────────────────────────────
PyMuPDF
python-docx
openpyxl

# ── Cache ─────────────────────────────────────────────────
chromadb

# ── Frontend ──────────────────────────────────────────────
dash
dash-bootstrap-components
